+++ CHANGELOG.md
@@ -55,3 +55,23 @@
 
 - Classes: `TelegramBot`, `UpdateController`
 - Functionality: user registration, CRUD operations with categories and notes, message and command handling
+
+---
+
+## [1.0.0] - 2024-08-14
+
+### Refactor
+
+#### Telegram bot
+
+- Refactored the bot's interaction mechanism with Telegram servers - replaced `TelegramLongPollingBot` with `TelegramWebhookBot`.
+
+#### Security
+
+- Refactored security configuration - for csrf the /callback/update path is included in the exceptions
+- also, /callback/update is marked as permission all
+
+### Added
+
+- **Controllers**:
+    - `TelegramWebhookController`